# i_ringer_was/i_ringer_was/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/users": {
      "post": {
        "description": "새로운 사용자를 생성합니다.",
        "operationId": "ModelsController_createUser",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "사용자 데이터",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "example": {
                  "role": "간호사",
                  "auth_id": "nurse001",
                  "password": "password123",
                  "nickname": "김간호사",
                  "hospital_id": "H001",
                  "ward_id": 1
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "사용자 생성",
        "tags": [
          "개별모델 삽입"
        ]
      }
    },
    "/api/patients": {
      "post": {
        "description": "새로운 환자를 생성합니다.",
        "operationId": "ModelsController_createPatient",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "환자 데이터",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "example": {
                  "nickname": "홍길동",
                  "chart_number": "P001"
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "환자 생성",
        "tags": [
          "개별모델 삽입"
        ]
      }
    },
    "/api/hospitals": {
      "post": {
        "description": "새로운 병원을 생성합니다.",
        "operationId": "ModelsController_createHospital",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "병원 데이터",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "example": {
                  "name": "서울대학교병원"
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "병원 생성",
        "tags": [
          "개별모델 삽입"
        ]
      }
    },
    "/api/wards": {
      "post": {
        "description": "새로운 병동을 생성합니다.",
        "operationId": "ModelsController_createWard",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "병동 데이터",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "example": {
                  "hospital_id": 1,
                  "name": "내과병동"
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "병동 생성",
        "tags": [
          "개별모델 삽입"
        ]
      }
    },
    "/api/rooms": {
      "post": {
        "description": "새로운 병실을 생성합니다.",
        "operationId": "ModelsController_createRoom",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "병실 데이터",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "example": {
                  "ward_id": 1,
                  "name": "101호",
                  "bed_count": 4
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "병실 생성",
        "tags": [
          "개별모델 삽입"
        ]
      }
    },
    "/api/beds": {
      "post": {
        "description": "새로운 침대를 생성합니다.",
        "operationId": "ModelsController_createBed",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "침대 데이터",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "example": {
                  "room_id": 1,
                  "bed_number": "A1",
                  "status": "available"
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "침대 생성",
        "tags": [
          "개별모델 삽입"
        ]
      }
    },
    "/api/devices": {
      "post": {
        "description": "새로운 링겔 모니터링 장치를 생성합니다.",
        "operationId": "ModelsController_createDevice",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "장치 데이터",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "example": {
                  "device_name": "I-Ringer-001",
                  "serial_number": "IR001",
                  "network_status": "online",
                  "battery_percent": 85,
                  "firmware_version": "1.0.0",
                  "bed_id": "1"
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "장치 생성",
        "tags": [
          "개별모델 삽입"
        ]
      }
    },
    "/api/patient_bed_assignments": {
      "post": {
        "description": "환자를 침대에 할당합니다.",
        "operationId": "ModelsController_createPatientBedAssignment",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "환자 침대 할당 데이터",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "example": {
                  "patient_id": 1,
                  "bed_id": 1,
                  "device_id": 1,
                  "infusion_type": "생리식염수",
                  "infusion_total_volumn": 500,
                  "assigned_at": "2025-09-04T14:00:00Z"
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "환자 침대 할당 생성",
        "tags": [
          "개별모델 삽입"
        ]
      }
    },
...
    "/api/notifications": {
      "post": {
        "description": "새로운 알림을 생성합니다.",
        "operationId": "ModelsController_createNotification",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "알림 데이터",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "example": {
                  "user_id": 1,
                  "patient_bed_assignment_id": 1,
                  "device_id": 1,
                  "title": "수액 주입 완료",
                  "message": "환자 홍길동의 수액 주입이 완료되었습니다.",
                  "type": "complete",
                  "is_read": 0
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "알림 생성",
        "tags": [
          "개별모델 삽입"
        ]
      }
    },
    "/api/user_settings": {
      "post": {
        "description": "새로운 사용자 설정을 생성합니다.",
        "operationId": "ModelsController_createUserSetting",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "사용자 설정 데이터",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "example": {
                  "user_id": 1,
                  "fast_enabled": 1,
                  "fast_threshold": 50,
                  "slow_enabled": 1,
                  "slow_threshold": -50,
                  "default_gatt": 60,
                  "complete_enabled": 1,
                  "complete_threshold": 95,
                  "stop_enabled": 1,
                  "alert_color": "#FF0000",
                  "alert_display_time": 5
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "사용자 설정 생성",
        "tags": [
          "개별모델 삽입"
        ]
      }
    },
    "/api/access_tokens": {
      "post": {
        "description": "새로운 액세스 토큰을 생성합니다.",
        "operationId": "ModelsController_createAccessToken",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "액세스 토큰 데이터",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "example": {
                  "user_id": 1,
                  "access_token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...",
                  "refresh_token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...",
                  "expires_at": "2025-09-04T15:00:00Z"
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "액세스 토큰 생성",
        "tags": [
          "개별모델 삽입"
        ]
      }
    },
...
  },
  "info": {
    "title": "I-Ringer API",
    "description": "Hospital IV drip monitoring system API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/",
      "description": "기본 서버"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}